initialize pandas series with float

+++ openbb_terminal/economy/econdb_model.py
@@ -503,10 +503,10 @@ def get_macro_data(
 
     if country not in COUNTRY_CODES:
         console.print(f"No data available for the country {country}.")
-        return pd.Series(), ""
+        return pd.Series(dtype=float), ""
     if parameter not in PARAMETERS:
         console.print(f"The parameter {parameter} is not an option for {country}.")
-        return pd.Series(), ""
+        return pd.Series(dtype=float), ""
 
     country_code = COUNTRY_CODES[country]
     country_currency = COUNTRY_CURRENCIES[country]
@@ -533,7 +533,7 @@ def get_macro_data(
                 f"No data available for {parameter} ({PARAMETERS[parameter]['name']}) "
                 f"of country {country.replace('_', ' ')}"
             )
-            return pd.Series(), ""
+            return pd.Series(dtype=float), ""
 
         if start_date or end_date:
             df = df.loc[start_date:end_date]
